separate out HTTP api server from the rest of the service
